refactor: :lipstick: Move link to source to about page only, tune spacing for fixed elements across browser and standalone
