+++ apps/www/src/styles/global.css.ts
@@ -76,7 +76,12 @@ globalStyle("#root", {
   isolation: "isolate",
   minHeight: "100vh",
   width: "100vw",
-  padding: `env(safe-area-inset-top) calc(env(safe-area-inset-right) + ${vars.space["2x"]}) env(safe-area-inset-bottom) calc(env(safe-area-inset-left) + ${vars.space["2x"]})`
+  padding: [
+    `calc(env(safe-area-inset-top) + ${vars.space["2x"]})`,
+    `calc(env(safe-area-inset-right) + ${vars.space["2x"]})`,
+    "env(safe-area-inset-bottom)",
+    `calc(env(safe-area-inset-left) + ${vars.space["2x"]})`
+  ].join(" ")
 })
 
 globalStyle("a, a:active", {
