+++ apps/www/src/views/about.tsx
@@ -1,5 +1,7 @@
 import { Meta, Title } from "solid-meta"
 
+import { SVGWrapper } from "../lib/svg-wrapper"
+
 import * as styles from "./about.css"
 
 function About() {
@@ -151,6 +153,31 @@ function About() {
           <a href="https://pnpm.io/">pnpm</a>
         </li>
       </ul>
+
+      <footer class={styles.appFooter}>
+        <a
+          class={styles.sourceLink}
+          aria-label="View source on GitHub"
+          href="https://github.com/chrstntdd/dicephrase"
+        >
+          <SVGWrapper>
+            <svg
+              class={styles.ghLogo}
+              clip-rule="evenodd"
+              fill-rule="evenodd"
+              stroke-linejoin="round"
+              stroke-miterlimit="2"
+              viewBox="0 0 136 133"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M67.9 0a67.9 67.9 0 0 0-21.5 132.3c3.4.6 4.6-1.5 4.6-3.3v-11.5c-19 4-22.9-9.1-22.9-9.1-3-7.9-7.5-10-7.5-10-6.2-4.2.4-4 .4-4 6.8.4 10.4 7 10.4 7 6 10.3 16 7.3 19.8 5.5.6-4.3 2.3-7.3 4.3-9-15-1.7-31-7.6-31-33.6 0-7.4 2.7-13.4 7-18.2a25 25 0 0 1 .7-18s5.7-1.8 18.7 7a65 65 0 0 1 34 0c13-8.8 18.6-7 18.6-7 3.7 9.4 1.4 16.3.7 18 4.3 4.8 7 10.8 7 18.2 0 26.1-15.9 31.8-31 33.5 2.4 2.1 4.6 6.3 4.6 12.6l-.1 18.6c0 1.8 1.2 4 4.7 3.3A67.9 67.9 0 0 0 67.9 0"
+                fill="currentColor"
+              />
+            </svg>
+          </SVGWrapper>
+        </a>
+      </footer>
     </section>
   )
 }
